Route cleanup
