+++ frontend/app/routes/protected/layout.tsx
